# Speed up graph view frontend by removing a querySelectorAll hot path