The query selector all inside the update links was slowing down production. This reduces that by storing the elements in an array.

100% performance increase from 480ms (2fps) to 225ms per frame (4fps) on a release build when dragging the sample document around.